idf/requirements.txt
fastapi==0.88.0
sqlalchemy==1.4.41
alembic==1.8.1
pydantic==1.9.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.5
uvicorn==0.20.0
python-dotenv==0.21.0
typing-extensions==4.4.0
